+++ delayed_image/delayed_nodes.py
@@ -11,12 +11,6 @@ from delayed_image import delayed_base
 from delayed_image import delayed_leafs
 from delayed_image.channel_spec import FusedChannelSpec
 
-
-try:
-    from line_profiler import profile
-except Exception:
-    from ubelt import identity as profile
-
 # --------
 # Stacking
 # --------
@@ -100,7 +94,6 @@ class ImageOpsMixin:
     if delayed_base.USE_SLOTS:
         __slots__ = tuple()
 
-    @profile
     def crop(self, space_slice=None, chan_idxs=None, clip=True, wrap=True,
              pad=0, lazy=False):
         """
@@ -259,7 +252,6 @@ class ImageOpsMixin:
             new = DelayedCrop(self, space_slice, chan_idxs)
         return new
 
-    @profile
     def _coordinate_crop(self, roi, lazy=False):
         """
         Experimental cropping implemented as a warp, which assumes the slice in
@@ -316,7 +308,6 @@ class ImageOpsMixin:
         # return self.warp(adjusted_transform, dsize=dsize, interpolation='linear')
         return self.warp(transform, dsize=dsize, interpolation='linear')
 
-    @profile
     def _padded_crop(self, space_slice, pad=0):
         """
         Does the type of padded crop we want, but inefficiently using a warp.
@@ -341,7 +332,6 @@ class ImageOpsMixin:
             new = new.warp(pad_warp, dsize=dsize)
         return new
 
-    @profile
     def warp(self, transform, dsize='auto', lazy=False, **warp_kwargs):
         """
         Applys an affine transformation to the image. See :class:`DelayedWarp`.
@@ -588,7 +578,6 @@ class DelayedChannelConcat(DelayedConcat, ImageOpsMixin):
     if delayed_base.USE_SLOTS:
         __slots__ = DelayedConcat.__slots__ + ('dsize', 'num_channels')
 
-    @profile
     def __init__(self, parts, dsize=None):
         """
         Args:
@@ -663,7 +652,6 @@ class DelayedChannelConcat(DelayedConcat, ImageOpsMixin):
             final = np.concatenate(stack, axis=2)
         return final
 
-    @profile
     def optimize(self):
         """
         Returns:
@@ -676,8 +664,7 @@ class DelayedChannelConcat(DelayedConcat, ImageOpsMixin):
             new._opt_logs.append('optimize DelayedChannelConcat')
         return new
 
-    @profile
-    def take_channels(self, channels):
+    def take_channels(self, channels, missing_channel_policy='return_nan'):
         """
         This method returns a subset of the vision data with only the
         specified bands / channels.
@@ -687,6 +674,12 @@ class DelayedChannelConcat(DelayedConcat, ImageOpsMixin):
                 List of integers indexes, a slice, or a channel spec, which is
                 typically a pipe (`|`) delimited list of channel codes. See
                 :class:`ChannelSpec` for more detials.
+
+            missing_channel_policy (str):
+                What to do if the requested channels are missing.
+                If set to 'return_nan' it will build a channel of nans which
+                will allow algorithms that can handle missing data to continue.
+                If set to 'error', then an ValueError will be raised.
 
         Returns:
             DelayedArray:
@@ -739,6 +732,9 @@ class DelayedChannelConcat(DelayedConcat, ImageOpsMixin):
             >>> from delayed_image.delayed_nodes import *  # NOQA
             >>> dset = kwcoco.CocoDataset.demo('vidshapes8-multispectral', use_cache=1, verbose=100)
             >>> self = delayed = dset.coco_image(1).delay()
+            >>> # by default requesting the channels that dont exist will
+            >>> # return nan channels, but this can be modified by setting
+            >>> # missing_channel_policy
             >>> channels = 'B1|foobar|bazbiz|B8'
             >>> new = self.take_channels(channels)
             >>> new_cropped = new.crop((slice(10, 200), slice(12, 350)))
@@ -747,12 +743,22 @@ class DelayedChannelConcat(DelayedConcat, ImageOpsMixin):
             >>> assert np.all(np.isnan(fused[..., 1:3]))
             >>> assert not np.any(np.isnan(fused[..., 0]))
             >>> assert not np.any(np.isnan(fused[..., 3]))
+            >>> # Test setting the missing channel policy
+            >>> import pytest
+            >>> with pytest.raises(ValueError):
+            >>>     new = self.take_channels(channels, missing_channel_policy='error')
+            >>> # test passing a bad policy
+            >>> with pytest.raises(KeyError):
+            >>>     new = self.take_channels(channels, missing_channel_policy='not-a-policy')
         """
         if channels is None:
             return self
         current_channels = self.channels
 
         if isinstance(channels, list):
+            # Allows channel selection via integer indexes
+            # TODO: this API needs to be tested and verified so it works
+            # correctly for integers indexes and string channel names.
             top_idx_mapping = channels
             top_codes = self.channels.as_list()
             request_codes = None
@@ -761,7 +767,7 @@ class DelayedChannelConcat(DelayedConcat, ImageOpsMixin):
             if current_channels == channels:
                 # If the request is equal to what we already have then skip this.
                 return self
-            # Computer subindex integer mapping
+            # Compute subindex integer mapping
             request_codes = channels.as_list()
             top_codes = current_channels.as_oset()
             top_idx_mapping = []
@@ -771,9 +777,28 @@ class DelayedChannelConcat(DelayedConcat, ImageOpsMixin):
                 except KeyError:
                     top_idx_mapping.append(None)
 
+        if missing_channel_policy == 'return_nan':
+            # default behavior
+            ...
+        elif missing_channel_policy == 'error':
+            # do a quick check for missing channels outside of the core logic
+            missing_channels = [code for idx, code in zip(top_idx_mapping, request_codes)
+                                if idx is None]
+            if missing_channels:
+                raise ValueError(
+                    f'Requested channels: {missing_channels} do not exist. '
+                    f'Available channels are: {current_channels}')
+        else:
+            raise KeyError(f'Invalid missing_channel_policy={missing_channel_policy}')
+
         # Rearange subcomponents into the specified channel representation
         # I am not confident that this logic is the best way to do this.
-        # This may be a bottleneck
+        # This may be a bottleneck.
+
+        # This object can contain multiple subimages, which each may have
+        # different numbers of channels. The FlatIndexer lets us pretend they
+        # are all flattened and map from the flat index to the nested index we
+        # can use to efficiently lookup the value.
         subindexer = kwarray.FlatIndexer([
             comp.num_channels for comp in self.parts])
 
@@ -996,7 +1021,6 @@ class DelayedImage(DelayedArray, ImageOpsMixin):
     if delayed_base.USE_SLOTS:
         __slots__ = DelayedArray.__slots__
 
-    @profile
     def __init__(self, subdata=None, dsize=None, channels=None):
         """
         Args:
@@ -1114,8 +1138,8 @@ class DelayedImage(DelayedArray, ImageOpsMixin):
         space_slice = (sl_y, sl_x)
         return self.crop(space_slice, chan_idxs)
 
-    @profile
-    def take_channels(self, channels, lazy=False):
+    def take_channels(self, channels, lazy=False,
+                      missing_channel_policy='return_nan'):
         """
         This method returns a subset of the vision data with only the
         specified bands / channels.
@@ -1124,11 +1148,17 @@ class DelayedImage(DelayedArray, ImageOpsMixin):
             channels (List[int] | slice | FusedChannelSpec):
                 List of integers indexes, a slice, or a channel spec, which is
                 typically a pipe (`|`) delimited list of channel codes. See
-                ChannelSpec for more detials.
+                :class:`ChannelSpec` for more detials.
 
             lazy (bool):
                 if True, dont create a new object if we can detect that it
                 would be a noop.
+
+            missing_channel_policy (str):
+                What to do if the requested channels are missing.
+                If set to 'return_nan' it will build a channel of nans which
+                will allow algorithms that can handle missing data to continue.
+                If set to 'error', then an ValueError will be raised.
 
         Returns:
             DelayedCrop:
@@ -1203,10 +1233,12 @@ class DelayedImage(DelayedArray, ImageOpsMixin):
                     for code in request_codes
                 ]
             except KeyError:
-                # If a requested channel doesn't exist we have to break this
-                # node up into a concat node with nans
+                # If a requested channel doesn't exist we break this node up
+                # into a concat node so we can use its nan handing logic
+                # This should be easy to optimize if necessary.
                 wrp = DelayedChannelConcat([self], dsize=self.dsize)
-                new =  wrp.take_channels(channels)
+                new =  wrp.take_channels(
+                    channels, missing_channel_policy=missing_channel_policy)
                 return new
         new_chan_ixs = top_idx_mapping
         new = self.crop(None, new_chan_ixs)
@@ -1262,7 +1294,6 @@ class DelayedImage(DelayedArray, ImageOpsMixin):
                                             channels=self.channels)
         return new
 
-    @profile
     def _opt_push_under_concat(self):
         """
         Push this node under its child node if it is a concatenation operation
@@ -1415,7 +1446,6 @@ class DelayedAsXarray(DelayedImage):
         final = xr.DataArray(subfinal, dims=('y', 'x', 'c'), coords=coords)
         return final
 
-    @profile
     def optimize(self):
         """
         Returns:
@@ -1445,7 +1475,6 @@ class DelayedWarp(DelayedImage):
     if delayed_base.USE_SLOTS:
         __slots__ = DelayedImage.__slots__ + ('_data_keys', '_algo_keys')
 
-    @profile
     def __init__(self, subdata, transform, dsize='auto', antialias=True,
                  interpolation='linear', border_value='auto', noop_eps=0,
                  backend='auto'):
@@ -1568,7 +1597,6 @@ class DelayedWarp(DelayedImage):
         final = kwarray.atleast_nd(final, 3, front=False)
         return final
 
-    @profile
     def optimize(self):
         """
         Returns:
@@ -1650,7 +1678,6 @@ class DelayedWarp(DelayedImage):
     def _transform_from_subdata(self):
         return self.transform
 
-    @profile
     def _opt_fuse_warps(self):
         """
         Combine two consecutive warps into a single operation.
@@ -1680,7 +1707,6 @@ class DelayedWarp(DelayedImage):
             new.print_graph()
         return new
 
-    @profile
     def _opt_absorb_overview(self):
         """
         Remove any deeper overviews that would be undone by this warp.
@@ -1918,7 +1944,6 @@ class DelayedWarp(DelayedImage):
             print('---------')
         return new
 
-    @profile
     def _opt_split_warp_overview(self):
         """
         Split this node into a warp and an overview if possible
@@ -2060,7 +2085,6 @@ class DelayedDequantize(DelayedImage):
             final = dequantize(final, quantization)
         return final
 
-    @profile
     def optimize(self):
         """
 
@@ -2095,7 +2119,6 @@ class DelayedDequantize(DelayedImage):
             new._opt_logs.append('optimize DelayedDequantize')
         return new
 
-    @profile
     def _opt_dequant_before_other(self):
         quantization = self.meta['quantization']
         new = copy.copy(self.subdata)
@@ -2144,7 +2167,6 @@ class DelayedCrop(DelayedImage):
     if delayed_base.USE_SLOTS:
         __slots__ = DelayedImage.__slots__
 
-    @profile
     def __init__(self, subdata, space_slice=None, chan_idxs=None):
         """
         Args:
@@ -2208,7 +2230,6 @@ class DelayedCrop(DelayedImage):
         self_from_subdata = kwimage.Affine.translate(offset)
         return self_from_subdata
 
-    @profile
     def optimize(self):
         """
         Returns:
@@ -2269,7 +2290,6 @@ class DelayedCrop(DelayedImage):
             new._opt_logs.append('optimize crop')
         return new
 
-    @profile
     def _opt_fuse_crops(self):
         """
         Combine two consecutive crops into a single operation.
@@ -2356,7 +2376,6 @@ class DelayedCrop(DelayedImage):
             new._opt_logs.append('Fuse crops')
         return new
 
-    @profile
     def _opt_warp_after_crop(self):
         """
         If the child node is a warp, move it after the crop.
@@ -2428,7 +2447,6 @@ class DelayedCrop(DelayedImage):
             new_outer._opt_logs.append('_opt_warp_after_crop')
         return new_outer
 
-    @profile
     def _opt_dequant_after_crop(self):
         # Swap order so dequantize is after the crop
         assert isinstance2(self.subdata, DelayedDequantize)
@@ -2537,7 +2555,6 @@ class DelayedOverview(DelayedImage):
         )
         return final
 
-    @profile
     def optimize(self):
         """
         Returns:
@@ -2569,7 +2586,6 @@ class DelayedOverview(DelayedImage):
         scale = 1 / 2 ** self.meta['overview']
         return kwimage.Affine.scale(scale)
 
-    @profile
     def _opt_overview_as_warp(self):
         """
         Sometimes it is beneficial to replace an overview with a warp as an
@@ -2583,7 +2599,6 @@ class DelayedOverview(DelayedImage):
             new._opt_logs.append('_opt_overview_as_warp')
         return new
 
-    @profile
     def _opt_crop_after_overview(self):
         """
         Given an outer overview and an inner crop, switch places. We want the
@@ -2652,7 +2667,6 @@ class DelayedOverview(DelayedImage):
             new._opt_logs.append('_opt_crop_after_overview')
         return new
 
-    @profile
     def _opt_fuse_overview(self):
         assert isinstance2(self.subdata, DelayedOverview)
         outer_overview = self.meta['overview']
@@ -2663,7 +2677,6 @@ class DelayedOverview(DelayedImage):
             new._opt_logs.append('_opt_fuse_overview')
         return new
 
-    @profile
     def _opt_dequant_after_overview(self):
         # Swap order so dequantize is after the crop
         assert isinstance2(self.subdata, DelayedDequantize)
@@ -2675,7 +2688,6 @@ class DelayedOverview(DelayedImage):
             new._opt_logs.append('_opt_dequant_after_overview')
         return new
 
-    @profile
     def _opt_warp_after_overview(self):
         """
         Given an warp followed by an overview, move the warp to the outer scope
@@ -2744,6 +2756,8 @@ class CoordinateCompatibilityError(ValueError):
 class _InnerAccumSegment:
     """
     Gather the indexes we need to take from an inner component
+
+    This is a helper for :func:`DelayedChannelConcat.take_channels`
     """
     if delayed_base.USE_SLOTS:
         __slots__ = ('comp', 'start', 'stop', 'codes', 'indexes')