+++ CHANGELOG.md
@@ -5,7 +5,24 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.4.2 - Unreleased
+## Version 0.4.3 - Unreleased
+
+### Added
+* Add `missing_channel_policy` to `.take_channels`.
+* Add `numel` to `SensorChanSpec` and `FusedSensorChanSpec`
+* Add `split` method to channel spec classes.
+* Add `from_spec` method to channel spec classes.
+
+### Fixed
+* `matching_sensor` now respects the special `*` sensor.
+* Corner cases when dsize is None / channels is None
+* `FusedSensorChanSpec.coerce` now works in simple cases.
+
+### Changed
+* Improved error messages
+
+
+## Version 0.4.2 - Released 2024-10-16
 
 ### Added
 * Can now work with different `warp_affine` backends.